+++ projects/dxilconv/lib/ShaderBinary/ShaderBinaryIncludes.h
@@ -18,7 +18,7 @@
 #include <d3d12.h>
 #define D3DX12_NO_STATE_OBJECT_HELPERS
 #include "dxc/Support/d3dx12.h"
-#include "dxc/D3D12TokenizedProgramFormat.hpp"
+#include "dxc/Support/D3D12TokenizedProgramFormat.hpp"
 #include "ShaderBinary/ShaderBinary.h"
 
 #define ASSUME( _exp ) { assert( _exp ); __analysis_assume( _exp ); __assume( _exp ); }
